Implement missing features of MAKE-SSL-*-STREAM
Finish implementation of MAKE-SSL-SERVER-STREAM to allow more
configuration when not reusing contexts.
CL+SSL only allows the setup of several configurations through the use of
contexts, therefore we use a disposable context to accommodate all
those configurations.

Implement MAKE-SSL-CLIENT-STREAM with all functionality described in
the documentation.

Refactor the creation of SSL streams as most of the code is shared
between client and server streams.

Tests Performed:
 - Build web-server image
 - Launch several configurations of the web-server: HTTP, HTTPS, and
 HTTPS with client authentication through certificate.
 - For each configuration perform a series of HTTP requests using: the
 browser, curl, and NET.ASERVE.CLIENT:DO-HTTP-REQUEST (both from ACL
 and SBCL).
 - Check for the appropriate behaviour forcing unusual
 situations (make an HTTPS request on a HTTP server, make an HTTP
 request on a HTTPS server, etc.)

Issue: PSCS-418
Reference: POA.37293.0
Change-Id: I29e1b96357187ea257506adcb527ca1ba03ec9f6
Reviewed-on: https://gerrit.siscog/c/open-source/zacl/+/6690
Reviewed-by: Rui Aurélio <rui.aurelio@siscog.pt>