Implement output buffering for ZACL-SIMPLE-STREAMs
Outputting to a ZACL-SIMPLE-STREAM (used in every connection with
clients) using FORMAT contains a giant overhead since it writes to the
response stream by small amounts.

To prevent excessively small outputting, this modification implements a
buffering mechanism that will improve the performance of writing to
ZACL-SIMPLE-STREAMs and therefore the performance of the HTTP server.

Issue: PSCS-603
Reference: POA.37293.0
Change-Id: I01c0f6321e212104a25e0c8d517fc378b82b042e
Reviewed-on: https://gerrit.siscog/c/open-source/zacl/+/7938
Reviewed-by: Luís Borges de Oliveira <lbo@siscog.pt>