Close the REAL-STREAM when closing a ZACL-SOCKET
While using HTTPS connections, a huge amount of alien heap space
wasn't being freed up.
Upon close analysis, it became apparent that this increase was
related to the creation of SSL-STREAMs.

Only by closing the REAL-STREAM of a ZACL-SOCKET (in HTTPS, this is a
SSL-STREAM), are the SSL related allocations freed from the alien heap
space, resolving the memory leak.

Issue: PSCS-1014
Reference: POA.37293.0
Change-Id: Ibfbba66cde479f49722bec5829e9556ee8d9fb38
Reviewed-on: https://gerrit.siscog/c/open-source/zacl/+/17915
Reviewed-by: Luís Borges de Oliveira <lbo@siscog.pt>