Implement TLS server name indication
CL+SSL supports the usage of SNI in TLS handshakes and all the
middleware ZACL implements also supports it, however the AllegroServe
Webserver limits its use on the availability of certain features on
the base Lisp.

With this we fake the existence of such features, allowing us to get
rid of yet another difference from the upstream AllegroServe.

Issue: PSCS-1208
Reference: POA.37293.0
Change-Id: I1be8885fa5e5f78d422c6199d5dbc8e97496c90a
Reviewed-on: https://gerrit.siscog/c/open-source/zacl/+/51773
Reviewed-by: Luís Borges de Oliveira <lbo@siscog.pt>
Reviewed-by: Vitor Santos <vitor.santos@siscog.pt>